Use uint64 for parseSize
Functions like RAMInBytes offen used to parse memory
set to cgroup, while Linux kernel take cgroup memory
value as uint64, so we need the return value of
RAMInBytes also be uint64, otherwise we are not able
to set memory larger than MAX(int64) and smaller than
MAX(uint64).

This'll somewhat break compatibility since parseSize
returns 0 when input is not parseable, but I doubt
people actually depend on that because we already
return error, so I don't think this'll be a problem.

Signed-off-by: Qiang Huang <h.huangqiang@huawei.com>